calibre 8.13.0
`calibre` is autobumped but the workflow failed to update to version
8.13.0 because `brew audit` failed with an "Artifact defined :monterey
as the minimum macOS version but the cask declared an on_system block
with a minimum macOS version of :ventura" error. This is a bug in
`Cask::Audit.cask_bundle_min_os`, as the main `Contents/Info.plist`
file in the app bundle isn't prioritized over `Info.plist` files in
other parts of the bundle. In this instance, the app declares Ventura
(13.3) as the minimum macOS version but `brew audit` is mistakenly
using the minimum version from a `QtWebEngineProcess.app` in
`QtWebEngineCore.framework` because the `Frameworks` directory comes
before `Info.plist` alphabetically.